include develop and recommends from META.json, not just MYMETA
Module::Build doesn't include prereqs listed in META when constructing
MYMETA.  And it doesn't have any options for specifying develop prereqs
or recommends other than meta_merge.  This makes it difficult to include
these prereqs in the MYMETA file.  The current MB and Dist::Zilla plugin
for it don't add any of these.  While that may be fixed in the future,
it will still be version limited to new MB versions.

Since the prereq list in MYMETA is most likely a superset of what is in
META, include the extra prereqs from META when installing.  In the case
where there are extra modules in META that shouldn't be in MYMETA, this
will just result in a possible extra module install attempt, but the
overall process will still continue.